This fix addresses 2 related situations. Both fixes are needed to help Ulauncher find .desktop folders in the same way as XDG spec defines it.
Fix 1:
Dedup the scanned .desktop files following the XDG spec where only
the .desktop file found in the first XDG_DATA_DIR paths is the one
we use. This is not happening currently and files located in the
$HOME/.local/share/applications were not found/prefered in some scenarios
where they file in /usr/share or other folder were either found first or
updated on user watch (i.e. upgrade of software)

Changes for this fix are only in util/desktop/reader.py

To clarify with an example:
Assuming XDG_DATA_DIR="/foo:/bar" and both /foo/firefox.desktop and
/bar/firefox.desktop exist, we should always pick /foo/firefox.desktop
to respect the XDG spec. However, if

Fix 2:
If two different .desktop files have the same Name= value only
one currently is shown in the search. However, per XDG spec the Name value
does not need to be unique, and only the .desktop file name found in first
XDG_DATA_DIR path should be (see Fix 1)

Changes for this fix are only in search/apps/AppdDb.py