Remove broken test that's not helpful anyway
All these actions should do is to send a message to the app via sockets
or dbus, and the app should handle the implemention. Now the
implementation and the test are handling internal logic of the window
and it doesn't even understand them and breaks without the useless
wrapper that was removed in 093c55c.