feat: Try to use wlr-layer-shell protocol on wayland
Uses https://github.com/wmww/gtk-layer-shell to set up the Ulauncher
window as a layer window, which drops decorations, allows positioning
and displaying over fullscreen windows.
Fails gracefully if the module cannot be loaded.
The extension is currently supported by wlroots based desktops (e.g Sway)
and KDE.
Moving the window to a different position is not supported in this mode
yet, but possibly could be implemented by changing the margin and or
anchor edges.