refactor: Pull out layer shell into its own class
Use separate `is_supported()` and `enable()` methods to allow clients
(i.e. descendants of `LayerShellOverlay`) to customize when to enable
the extension.
The specifics of focus handling / positioning are not exposed to clients,
this can be done later when other uses of the protocol are added.
(One such use would to show which edge the window is anchored to during
positioning when it is implemented in this mode.)