fix: Fix dragging when clicking on results
The event coordinates are relative to the widget being clicked on instead
of the window. Therefore it was possible to activate dragging by clicking
on the results.
Add an EventBox widget around the draggable area and subscribe to the
click event only on that instead of the whole window in order to fix it.