fix: Rewrite ExtensionDownloader to use git repo discover references
The official API have very strict rate limits

The discover references is what the git client uses internally to fetch
from the server, so it's like using git but without the extra
dependency.

Unfortunately you can't get timestamps for the commits this way. Just
the references and hashes. To get timestamps you need to make a full
commit.

This currently breaks the tests and removes a lot of error handling.
Most of that I doubt we can/should restore without fetching the extra
metadate.

It's possible to use te old way to fetch only the date when we know
there is an update available and hence it would avoid the rate limits,
but I think it doesn't justify adding all that old complexity back.