chore: add back last_commit_time
This assumes the mtime of the root directory in the archived tar file is the commit date,
I believe I tested this for GitHub, Codeberg and GitLab, but even if
it's not true it would probably be more reliable than nothing.

We don't actually use it any more because we catch it too late for it to
be meaningful, but if we revert this change it's better to keep it.