docs: update extension debugging instructions
I think this is already a big improvement from before when you had to
manually copy or symlink your local repo, or run it with the weird long
CLI instructions.

Maybe we want a command on top of this to simply run your extension
without installing it, but I think it's mostly over-thinking and a
bad idea. If extensions are allowed to run from any path we need to keep
track of that when we load/handle/change the extension settings for example.
We normally use the extension id (derived from the URL) as the unique
identifier to handle the settings and other things.