docs: remove debug instructions
Someone reported that this doesn't work. I don't have the time to test but I think it was overly complex and a bad idea to describe how to debug extensions in this way.
We are removing this anyway in Ulauncher v6 and I don't want to be bothered with issues about it, so removing the instructions

You never needed this ability.

fixes #1328