refactor: move shortcut trigger to its own result
To avoid confusion and spaghetti code we want the Shortcut trigger
separate from the result (too many cases to test/maintain for the same
result class before), making it hard to read and too easy to break.

The trigger now handles expanding the search term to the keyword
and activating without arguments, while the result handles/expects an
argument (although it can be empty)

Additionally, moving the description getter to the shortcut mode to set
a static description allows further simplification later on.