refactor: remove now unused result.get_description
As of the previous commit this is completely redundant and all
descriptions are static.

One possible reason to leave it as deprecated instead is that extensions
might use it, but that would not be useful (they set the description, so
they don't need to get it. I also searched all GitHub repos with "Ulauncher"
and "get_description", and no public extensions on GitHub does use this.