chore: ignore system packages that satisfy the required dev dependencies (#1463)
If you have these packages installed on your system pip tries to use those versions.
This has the potential to break our scripts since a global mypy for example won't
work with the pygobject stubs installed for the .venv

This does make it a few seconds slower to re-run "make python-venv" on my system.