perf: defer loading mode handler until after UI has loaded (#1493)
* perf: defer loading mode handler until after UI has loaded

Mode handler contains all the stuff that's slow to load. I did previously deferr loading it, but I broke this in recent changes (or maybe in less recent changes too).

Additionally, the event I previously used to load the deferred logic now runs before the window is shows and blocks it from showing. I'm not sure if GTK/Gnome made changes to gazlight me or if this never really worked.

Also refactored the deferring logic to be more self-contained in the query handler.

* test: fix broken test