refactor: clean up get_most_frequent exception
This still doesn't feel like an optimal solution, and it still has some
hard coded logic/semantics. The underlying problem is that we have made
assumptions in the design that the default results is the top applications
rather than making that configurable. I would prefer if the architecture
would allow either all or none of the modes to populate this.