refactor(remote): rewrite remore and url parser for better separation
makes the parse method and result more self contained, ie: handle all
parsing and the host, path and id in one place, so the parse result
doesn't have to expose internals.

Adds a new "download_url_template" prop so we don't need to
conditionally check the host outside of the parser.

Removes the "use_git" prop, which didn't belong in the url parser
anyway, and now we can check if we have git and not the
"download_url_template" prop instead.

Also replaced TypeDict with BaseDataClass as the baseclass for UrlParseResult,
Then used UrlParseResult as the baseclass for ExtensionRemote (avoids
some repetitious code).

Rewrote the tests to account for the new changes, to simplify, and to
be more isolated (only check the props that has to do with the test).
Otherwise changing any prop will break all the tests.

Added some new tests to cover that file:// protocol or not get the same
resaul get the same resaults, and similar.