perf: defer expensive window operations (#1513)
* refactor(window): reorganize initialization logic

* perf: make ulauncher handle keyboard input before fully ready

1. Sets the window to be invisible (opacity=0) initially
2. Moves all the expensive/slow styling and positioning logic to a new method "apply_styling", which also removes the opacity=0
3. If not is_composited, run apply_styling() before showing the window (because the opacity=0 won't work), otherwise wait until after the initial render.

In my profiling it now takes about 65ms for the window to render and be able to get the keyboard input (35% speed improvement). This seems to be the last improvement we can do. The rest is all unavoidable GTK initlialization stuff.

(note: I removed the migration and kill_ulauncher_v5 scripts in my performance testing, those will be removed anyway soon after v6 is released)