feat: basic CLI for extension management (#1508)
* feature #1150: basic CLI for extension management

* chore #1150: improvements for extension CLI

* fix(cli): use info loglevel for cli mode unless verbose flag

* refactor(cli): create logger for whole module

At least in my tests, the logging module is very fast to set up, so I
see no point to defer loading it.

* fix: add better message for extension already being on the latest version

There was just a message that it was "checking", and no update on the
result.

We also don't want the CLI to say the former.

* refactor: rename all input arguments to 'input'

This makes it easier to redirect to a different method. Ex if you use
the install command but the extension already exists.

Also, I find it that lowercase looks more standardized in the --help command output.

* fix: rewrite cli type parsing logic

This fixes so that urls starting with "file://" will work again, and
that the path resolve logic applies for both install and update.

* feat: trigger extension restart in the main app

* fix: also handle stopping extension after uninstall

* chore: made flat structure for extension CLI commands

---------

Co-authored-by: Albin Larsson <515120+friday@users.noreply.github.com>