chore: add/fix types for tests
The fixture argument (unless it's another fixture) is MockerFixture, and
the type when you use the fixture is MagicMock, but Mypy doesn't let you
use that if you return the ".return_value" prop.

Overall the MockerFixture type is helpful, but MagickMock is about as
useful as Any. If not using the special assert method you are getting
more helpful results by setting the type to be the mocked objects type,
although it really isn't.