perf: index trigger keywords instead of parsing query through modes
core already has all the triggers, so we can just make sure the triggers have shortcuts defined (if they are triggers for shortcuts, eg not app results), then make an additional dict storing (with keywords as key).

This is faster and smarter in multiple ways than "parsing" the query. Now instead the old parse method will only apply to check if the query is a match, and it will only apply for modes without query keywords or with optional query keywords.

Now "matches_query_str" is only for modes that aren't expecting keywords