refactor(dbus): clarify which dbus methods are general vs ulauncher app specific
The new is_app_running method intentionally takes an argument so we can potentially use
it to check other apps or other versions of Ulauncher (we could use this
to improve kill_ulauncher_v5() or potentially the launch_app command
(probably not though).

The older ones are specifically for communicating with the Ulauncher app
under the current app id.

Also removed exception handling because we should always have Gio and if
it raises an error we don't want to default to anything.