refactor(ext): fix ext mode handle_query return type inconsistency
handle_query is expected to return the results, but ext mode is async,
so we don't have them yet, and instead used the Any return type and
returned True to keep the window open.

This was pretty confusing and non-explicit, so I changed the return type
for handle_query to also allow None to do nothing (until we have the
data).

Could also simplify some socket controller internals as a result of the
change.

And removed a redundant test