fix(cli): simplify the parsing, ext controller interaction and improve error handling
leverages parse_extension_url to make a simpler method to get the
controller, instead of get_argument_type(). Also change the meaning of
the controller getter method to only return for installed extensions,
which simplifies our error handling, and aligns better with how we've
discussed the ExtController should behave in the near future.

The old error handling was using a lot of try/except with bare excepts,
overly verbose errors