fix(ext-controller): always remove the state after removing extension
The state should not be preserved when fallbacking to a non-managed
extension. The state mostly reflect the last update and details to be
able to update it that doesn't apply to non-managed extensions. The
error also should be reset since it's not the same location/version and
we need to reset the error to be able to start it from the new location.

Then the only properties left is is_enabled which we already override,
and the id which is recreated when you in the ext controller
constructor.