refactor: make runtime.stop non async
I ran into issues trying to write tests for the runtime because it was
async in a different branch.

This reuses the existing callback instead of adding a new one, and in
addition it replaces await asyncio.sleep with a more sane timer. That
way we only wait as long as necessary (which is still 0.5s, but due to
the Client implementation instead)