chore: make pyrefly require type annotations
Pyrefly can infer type annotations, which is great, but I think more so
this is something good for untyped external libs. Our own code is
strictly typed now and I prefer to keep it that way.

This prevents the important half of the issues in afc8767f. The other problem is the lack if None return type. I don't know if pyrefly has that rule.