fix(core): rewrite cache to be per mode level
Adds a flag for modes that have changed keywords/triggers and needs to be reloaded in core. This is only really for extensions. All other modes load fast and synchronously

If we add support for running extensions on demand instead of in the background, we will still need this for the preview extension logic.

Then changes the caches to per mode level. This fixes so that we can reload extensions when they have changed, without having to reload apps etc for every keypress.