refactor: avoid using the same eventbus prefix as for extension mode
The previous use of the eventbus here used a variable name and passed
that to the registry, as which made it seem like it was an isolated
comms channel, but it's not how these works. They operate by prefixes
and this used the same prefix as extension mode already has.