ci: ignore broken mypy rule
call-arg is broken if a sub class overrides the type definition

We use pyrefly instead of mypy now with mypy just left in the ci in case
pyrefly would miss something (it's still beta), so ignoring this broken
rule is ok.

We are also stuck on an old mypy release because we want to support
older python versions.