refactor: remove misleading result callback handling (#1590)
This was a remainder from the v5 implementation details. It confused me because there is no way for the extension result to still have callback after going through JSON serialization. No one has complained about extensions breaking in v6, even though we probably switched to use JSON a couple of years ago.

Then I looked into it:
In v5, the ResultItem class itself had an "on_enter" method. Our internal result items like CalcResultItem etc used that directly. The ExtensionResultItem took an `on_enter` constructor param and the constructor stored it as `_on_enter`. Then `on_enter()` for extension results just returned self._on_enter. So in effect this was a callback then, but only because of the internal implementation details, which we have changed (for the better).