refactor: move query debounce to the extension
This is arguably also a fix: We use the incoming input (query) change events to invalidate outdated responses, but when we also debounce from the app, it means this information is not current, so we send responses unnecessaily.

One could argue that this change means we instead send events from the app to the extension unnecessarily, which is true. But I do think it makes a more consistent control flow for both sides if we send all event types the same way and can rely on them to be recieved in time.

Additionally, we could use this information to cancel ongoing threads (I just think that needs more refactoring first).