chore: rename make check -> make lint, and make test -> make check
Having the command for running all checks named "test" was confusing
imo. "test" is easy to confuse with (unit) tests.

Now there is no `make test`, which is intentional. If someone runs it
they will realize it's gone and they have to check the makefile or docs
again.

In the future I think it makes sense to rename `make pytest` into `make test`